remove and relocate components
