+++ plugins/performance-resources/src/components/kra/KRAAssigneesPresenter.svelte
@@ -41,6 +41,7 @@
   $: persons = filter(items.map((i) => $personIdByAccountId.get(i.employee)))
     .map((p) => $personByIdStore.get(p))
     .filter((p) => p !== undefined)
+    .map((p) => p as Person)
 </script>
 
 <div
